addressing PR feedback: rename "standard" to "default" AWS signer; support hashing partial arrays; more tests and logging for intermediate signing steps; rework body hash classes
