+++ runtime/auth/aws-signing-common/common/src/aws/smithy/kotlin/runtime/auth/awssigning/AwsSigningConfig.kt
@@ -51,41 +51,6 @@ enum class AwsSignatureType {
     HTTP_REQUEST_EVENT,
 }
 
-/**
- * Identifies a source for calculating the body hash value
- */
-sealed class BodyHash(open val hash: String?) {
-    /**
-     * The hash value should be calculated from the body payload
-     */
-    object CalculateFromPayload : BodyHash(null)
-
-    /**
-     * The hash value should indicate an unsigned payload
-     */
-    object UnsignedPayload : BodyHash("UNSIGNED-PAYLOAD")
-
-    /**
-     * The hash value should indicate an empty body
-     */
-    object EmptyBody : BodyHash("e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855") // hash of ""
-
-    /**
-     * The hash value should indicate that signature covers only headers and that there is no payload
-     */
-    object StreamingAws4HmacSha256Payload : BodyHash("STREAMING-AWS4-HMAC-SHA256-PAYLOAD")
-
-    /**
-     * The hash value should indicate ???
-     */
-    object StreamingAws4HmacSha256Events : BodyHash("STREAMING-AWS4-HMAC-SHA256-EVENTS")
-
-    /**
-     * Use an explicit, precalculated value for the hash
-     */
-    data class Precalculated(override val hash: String) : BodyHash(hash)
-}
-
 enum class AwsSignedBodyHeader {
     /**
      * Do not add a header.
@@ -168,10 +133,10 @@ class AwsSigningConfig(builder: Builder) {
 
     /**
      * Determines the source of the canonical request's body public value. The default is
-     * [BodyHash.CalculateFromPayload], indicating that a public value will be calculated from the payload during
-     * signing.
+     * [HashSpecification.CalculateFromPayload], indicating that a public value will be calculated from the payload
+     * during signing.
      */
-    val bodyHash: BodyHash = builder.bodyHash ?: BodyHash.CalculateFromPayload
+    val hashSpecification: HashSpecification = builder.hashSpecification ?: HashSpecification.CalculateFromPayload
 
     /**
      * Determines which body "hash" header, if any, should be added to the canonical request and the signed request.
@@ -203,7 +168,7 @@ class AwsSigningConfig(builder: Builder) {
         it.useDoubleUriEncode = useDoubleUriEncode
         it.normalizeUriPath = normalizeUriPath
         it.omitSessionToken = omitSessionToken
-        it.bodyHash = bodyHash
+        it.hashSpecification = hashSpecification
         it.signedBodyHeader = signedBodyHeader
         it.credentialsProvider = credentialsProvider
         it.expiresAfter = expiresAfter
@@ -219,7 +184,7 @@ class AwsSigningConfig(builder: Builder) {
         var useDoubleUriEncode: Boolean = true
         var normalizeUriPath: Boolean = true
         var omitSessionToken: Boolean = false
-        var bodyHash: BodyHash? = null
+        var hashSpecification: HashSpecification? = null
         var signedBodyHeader: AwsSignedBodyHeader = AwsSignedBodyHeader.NONE
         var credentialsProvider: CredentialsProvider? = null
         var expiresAfter: Duration? = null
