+++ runtime/auth/aws-signing-tests/common/src/aws/smithy/kotlin/runtime/auth/awssigning/tests/BasicSigningTestBase.kt
@@ -118,7 +118,7 @@ abstract class BasicSigningTestBase : HasSigner {
         useDoubleUriEncode = false
         normalizeUriPath = true
         signedBodyHeader = AwsSignedBodyHeader.X_AMZ_CONTENT_SHA256
-        bodyHash = BodyHash.StreamingAws4HmacSha256Payload
+        hashSpecification = HashSpecification.StreamingAws4HmacSha256Payload
         credentialsProvider = CHUNKED_TEST_CREDENTIALS_PROVIDER
     }
