+++ runtime/hashing/jvm/src/aws/smithy/kotlin/runtime/hashing/Md5JVM.kt
@@ -8,7 +8,7 @@ import java.security.MessageDigest
 
 actual class Md5 : Md5Base() {
     private val md = MessageDigest.getInstance("MD5")
-    override fun update(input: ByteArray) = md.update(input)
+    override fun update(input: ByteArray, offset: Int, length: Int) = md.update(input, offset, length)
     override fun digest(): ByteArray = md.digest()
     override fun reset() = md.reset()
 }
