+++ runtime/hashing/jvm/src/aws/smithy/kotlin/runtime/hashing/Sha256JVM.kt
@@ -8,7 +8,7 @@ import java.security.MessageDigest
 
 actual class Sha256 : Sha256Base() {
     private val md = MessageDigest.getInstance("SHA-256")
-    override fun update(input: ByteArray) = md.update(input)
+    override fun update(input: ByteArray, offset: Int, length: Int) = md.update(input, offset, length)
     override fun digest(): ByteArray = md.digest()
     override fun reset() = md.reset()
 }
