+++ settings.gradle.kts
@@ -35,7 +35,7 @@ include(":runtime")
 include(":runtime:auth:aws-credentials")
 include(":runtime:auth:aws-signing-common")
 include(":runtime:auth:aws-signing-crt")
-include(":runtime:auth:aws-signing-standard")
+include(":runtime:auth:aws-signing-default")
 include(":runtime:auth:aws-signing-tests")
 include(":runtime:crt-util")
 include(":runtime:hashing")
